LPS-44643 Breadcrumb classes moved back to portlet boundaries. This is needed because otherwise, the kernel will depend on breadcrumb portlet.
