+++ portal-web/docroot/html/portlet/breadcrumb/init.jsp
@@ -16,8 +16,8 @@
 
 <%@ include file="/html/portlet/init.jsp" %>
 
-<%@ page import="com.liferay.portal.kernel.servlet.taglib.ui.BreadcrumbEntry" %><%@
-page import="com.liferay.portlet.breadcrumb.util.BreadcrumbUtil" %>
+<%@ page import="com.liferay.portlet.breadcrumb.BreadcrumbEntry" %><%@
+page import="com.liferay.portlet.breadcrumb.BreadcrumbUtil" %>
 
 <%
 String portletResource = ParamUtil.getString(request, "portletResource");
